Move blocks/MoviesList to components
