+++ src/pages/index.js
@@ -9,7 +9,7 @@ import { ClientProvider } from 'context/ClientContext'
 import get from 'utils/get'
 import Header from 'components/Header'
 import Filters from 'blocks/Filters'
-import MoviesList from 'blocks/MoviesList/index'
+import MoviesList from 'components/MoviesList/index'
 import { LANGUAGES } from 'data/constants'
 import { LanguageProvider } from 'context/LanguageContext'
 import useLocalStorage from 'hooks/useLocalStorage'
